refactor: 추후 type 병합을 위해 파일 이동 및 API URL 수정
