+++ src/pages/RegisterPage/types.ts
@@ -1,3 +1,4 @@
+//삭제 예정
 export interface Post {
   likes: Like[];
   comments: Comment[];
@@ -60,17 +61,8 @@ export interface User {
   posts: Post[];
   likes: Like[];
   comments: string[];
-  followers: [];
-  following: [
-    {
-      _id: '6169e91316cb2265df003c6d';
-      user: '6169e58216cb2265df003bf4';
-      follower: '6169e206aa57d952c6dc1edd';
-      createdAt: '2021-10-15T20:48:19.816Z';
-      updatedAt: '2021-10-15T20:48:19.816Z';
-      __v: 0;
-    },
-  ];
+  followers: string[];
+  following: string[];
   notifications: Notification[];
   messages: Message[];
   _id: string;
